Use "long" Type for System Call Result, and Print Linux Task ID (TID) As "unsigned long".

+++ as/src/base/as.c
@@ -304,7 +304,7 @@ main(int argc, char **argv)
 #ifdef USE_ASM
 	g_asm_hook_enabled = g_asm_cb_enabled = c->asmalloc_enabled;
 
-	int initial_tid = syscall(SYS_gettid);
+	long initial_tid = syscall(SYS_gettid);
 #endif
 
 #ifdef MEM_COUNT
@@ -356,10 +356,10 @@ main(int argc, char **argv)
 
 #ifdef USE_ASM
 	// Log the main thread's Linux Task ID (pre- and post-fork) to the console.
-	fprintf(stderr, "Initial main thread tid: %d\n", initial_tid);
+	fprintf(stderr, "Initial main thread tid: %lu\n", initial_tid);
 
 	if (! run_in_foreground && c->run_as_daemon) {
-		fprintf(stderr, "Post-daemonize main thread tid: %d\n",
+		fprintf(stderr, "Post-daemonize main thread tid: %lu\n",
 				syscall(SYS_gettid));
 	}
 #endif
